Fixed issue with escaped \\ symbols
Depending on whether parameters were passed from command line or not, ParametersTask was executed or not executed at all. As it reads and un-escapes \{, \} and \\ sequences - this behavior led to very non-intuitive behavior: if you pass parameters you have to double your \ symbols (so \\ should become \\\\), if you don't - they are not processed (so \\ in transform xml stays \\ in the final file or attribute). This change makes ParametersTask execution non-optional, so you get consistent behavior whether you passed parameters or not. \ is treated like escape char by ctt.exe, so you have to double it if it participates in \{, \} or \\ sequences that are not escapements of parameters.